added a plotly_prev method to the  ipyrun._displayfile.DisplayFile class mapping to display .plotly files.
idea being that when we make plotly widgets we can export them both as png's but also .plotly's so the user will get the benefits of the interactive stuff when viewing the output... 

note.
-----
.plotly files are simply .json files and can be treated as such. (i.e. json.load(fpth))
it would be better to save them as .plotly.json files but then they are treated by .json files from the DisplayFile class.